+++ LeisBaTibo/Data Model/Data.swift
@@ -7,7 +7,7 @@ class Data: Object {
     @objc dynamic var brand: String = ""
     @objc dynamic var store: String = ""
     @objc dynamic var buyingQuantity: Double = 0
-    @objc dynamic var estimatedPrice: Float = 0.0
+    @objc dynamic var estimatedPrice: Double = 0
     @objc dynamic var unitMeasurement: String = ""
     @objc dynamic var quantityForEstimatedPrice: Double = 0
     @objc dynamic var isiJikaUnitBag: Double = 0
@@ -15,7 +15,7 @@ class Data: Object {
  
     
     @objc dynamic var subCategory: String = ""
-    @objc dynamic var salePrice: Float = 0.0
+    @objc dynamic var salePrice: Double = 0.0
     //@objc dynamic var picture: CIImageOption? = nil
